==> gcp_files/Step_4_final_model_run-HL_gcp_dataproc.ipynb <==
# w261 Final Project - Clickthrough Rate Prediction


Team 2  
Danielle Adler, Craig Fujii, Conor Healy, YoungKoung Kim

Summer 2019, section [Your section numbers>]   

## Setup

# imports
import re
import ast
import time
import shutil
import os
import copy
import numpy as np
import seaborn as sns
import networkx as nx
import matplotlib.pyplot as plt
from IPython.display import display

from pyspark.sql import Window, Row
from pyspark.sql.functions import col, desc, mean, isnan, when, count, isnull, rank, sum, countDistinct, avg, stddev, round, lit, rand, broadcast, udf, log, monotonically_increasing_id
from pyspark.sql.types import LongType, IntegerType, StringType, DoubleType, ArrayType, FloatType
from pyspark.sql.window import Window

from pyspark.ml import Pipeline
from pyspark.ml.feature import StringIndexer, OneHotEncoder, OneHotEncoderEstimator, VectorAssembler, MinMaxScaler, Imputer
from pyspark.ml.stat import Correlation
from pyspark.mllib.stat import Statistics

from pyspark.ml.linalg import Vectors
from pyspark.ml.classification import LogisticRegression, DecisionTreeClassifier
from pyspark.ml.evaluation import MulticlassClassificationEvaluator
from pyspark.ml.tuning import CrossValidator, ParamGridBuilder

import warnings
warnings.filterwarnings('ignore')

%reload_ext autoreload
%autoreload 2

# store path to notebook
PWD = !pwd
PWD = PWD[0]

# # start Spark Session
# from pyspark.sql import SparkSession
# app_name = "final_project"
# master = "local[*]"
# spark = SparkSession\
#         .builder\
#         .appName(app_name)\
#         .master(master)\
#         .getOrCreate()
# sc = spark.sparkContext
spark

sc = spark.sparkContext

master_start = time.time()

# __Section 4__ - Algorithm Implementation

## Functions

def metrics(col_name):
    """Calculates key metrics on a binary column of 0s and 1s
    Input: Column Name
    Output: TP, FP, TN, FP, accuracy, precision, recall, F1 score
    """
    
    TP = train_set.filter(col('_c0')==0).filter(col(col_name)==0).count()
    FP = train_set.filter(col('_c0')==1).filter(col(col_name)==0).count()
    TN = train_set.filter(col('_c0')==1).filter(col(col_name)==1).count()
    FN = train_set.filter(col('_c0')==0).filter(col(col_name)==1).count()
    accuracy = (TP + TN) / (TP + FP + FN + TN)
    precision = TP/(TP+FP)
    recall = TP/(TP+FN)
    F1 = 2*(precision*recall)/(precision+recall)
    print("True Positives: {} \nFalse Positives: {} \nTrue Negatives: {} \nFalse Negatives: {} \nAccuracy: {} \nPrecision: {} \nRecall: {} \nF1 Score: {}".format(TP, FP, TN, FN, accuracy, precision, recall, F1))

def feature_prep (input_df):
    """Parses out types of features within a dataframe
    Input: Dataframe
    Output: Dataframe, Categorical Column List, Numeric Column List
    """
    
    # develop feature set
    dep_var = ['_c0']

    model_df = input_df.select(input_df.columns[1:])

    feature_df = model_df.select([column for column in model_df.columns 
                              if column not in dep_var])
    
    # numerical and categorical column split
    cat_cols = feature_df.columns[1:]
    num_cols = feature_df.columns[0]
    
    return model_df, cat_cols, num_cols

def column_vector_maker(df, columns, lookup_df, suffix):
    """One-hot encodes categorical variables
    Input: Dataframe, Categorical Column List, Lookup Dataframe, Suffix for New Columns
    Output: Dataframe, Categorical Column List, Numeric Column List
    """
    
    # Creating the lookup table vectors for each category
    # we are doing a map-side join
    for i in columns:
        new_col = str(i) + str(suffix)
        df = df.join(lookup_df, df[i]==lookup_df['id'], "left_outer" )
        df = df.withColumnRenamed("category", new_col) 
        lst = df.columns
        cols = [i for i in lst if i not in ['id']]
        df = df.select(cols)
        
    df = df.select([c for c in df.columns if c not in columns])
    
    cat_cols = df.columns[2:]
    num_cols = df.columns[0]
    
    return df, cat_cols, num_cols

def pipeline_prep (object_type):
    """Assembles a dataframe of label and feature fectors for machine learning algorithm
    Input: Dataframe, Categorical Column List, Numeric Column List
    Ouput: Dataframe
    """
    
    model_df = object_type[0]
    cat_cols = object_type[1]
    num_cols = object_type[2]

    # developing the string indexer and vector assembler of input and output columns
    label_stringIdx = StringIndexer(inputCol = '_c0', outputCol = 'label')

    assemblerInputs = cat_cols + [num_cols]
    assembler = VectorAssembler(inputCols=assemblerInputs, outputCol="features")
    
    # pipeline work for more setup
    pipeline = Pipeline(stages = [label_stringIdx, assembler])

    pipelineModel = pipeline.fit(model_df)
    model_df = pipelineModel.transform(model_df)

    selectedCols = ['label', 'features'] 
    model_df = model_df.select(selectedCols)
    return model_df

def lr_cross_validate (lr_model):
    """Hypertunes a logistic regression model based on a set of parameters
    Input: Logistic regression classification
    Output: Hypertuned logistic regression classification
    """

    # for elasticNetParam, 0 is L2 and 1 is L1
    paramGrid = ParamGridBuilder() \
        .addGrid(lr_model.elasticNetParam, [0.0, 0.25, 0.75, 1.0]) \
        .addGrid(lr_model.regParam, [0.0, 0.01, 0.1, 0.5, 0.75, 1.0, 1.25]) \
        .build()

    # optimizing for an f1 score
    crossval = CrossValidator(estimator = lr, estimatorParamMaps = paramGrid,
                              evaluator = MulticlassClassificationEvaluator(
                                  labelCol = "label", predictionCol="prediction", 
                                  metricName="f1"), numFolds=3)
    return crossval

def dt_cross_validate (dt_model):
    """Hypertunes a decision tree model based on a set of parameters
    Input: Decision tree classification
    Output: Hypertuned decision tree classification
    """

    paramGrid = ParamGridBuilder() \
        .addGrid(dt_model.maxBins, [28, 30, 32, 34, 36]) \
        .addGrid(dt_model.maxDepth, [3, 4, 5, 6, 7]) \
        .addGrid(dt_model.impurity, ['gini', 'entropy']) \
        .build()

    # optimizing for an f1 score
    crossval = CrossValidator(estimator = dt, estimatorParamMaps = paramGrid,
                              evaluator = MulticlassClassificationEvaluator(
                                  labelCol = "label", predictionCol="prediction", 
                                  metricName="f1"), numFolds=3)
    return crossval

def train_results (trainingSummary):
    """Showcases key metrics on logistic regression training summary
    Input: Logistic regression training model
    Output: Training set accuracy, precision, recall, F1 score
    """

    # predicting on the training set
    falsePositiveRate = trainingSummary.weightedFalsePositiveRate
    truePositiveRate = trainingSummary.weightedTruePositiveRate
    accuracy = trainingSummary.accuracy
    precision = trainingSummary.weightedPrecision
    recall = trainingSummary.weightedRecall
    f1 = trainingSummary.weightedFMeasure()

    print("False Positive Rate: %s\nTrue Positive Rate: %s\nAccuracy: %s\nPrecision: %s\nRecall: %s\nF1 Score: %s"
          % (falsePositiveRate, truePositiveRate, accuracy, precision, recall, f1))

def test_results (predictions):
    """Showcases key metrics on all algorithm test datasets
    Input: Classification training model
    Output: Testing set accuracy, precision, recall, F1 score
    """
    
    evaluator = MulticlassClassificationEvaluator(labelCol = "label", predictionCol="prediction", metricName="accuracy")
    print("Accuracy:", evaluator.evaluate(predictions))

    evaluator = MulticlassClassificationEvaluator(labelCol = "label", predictionCol="prediction", metricName="weightedPrecision")
    print("Precision:", evaluator.evaluate(predictions))
    
    evaluator = MulticlassClassificationEvaluator(labelCol = "label", predictionCol="prediction", metricName="weightedRecall")
    print("Recall:", evaluator.evaluate(predictions))
    
    evaluator = MulticlassClassificationEvaluator(labelCol = "label", predictionCol="prediction", metricName="f1")
    print("F1 Score:", evaluator.evaluate(predictions))

def to_array(col):
    """Separating a vector into distinct columns
    Input: Vector column
    Output: Separate array columns of vector
    """
    
    # separating a vector into its array types
    def to_array_(v):
        return v.toArray().tolist()
    return udf(to_array_, ArrayType(DoubleType())).asNondeterministic()(col)

def null_as_Missing(x):
    return when(col(x).isNotNull(), col(x)).otherwise('L')

<br>

## Reading in Model Files

# read parquet files of features
start = time.time()

folder_hl = 'data_good'
# mean_wgt_HL_df = spark.read.parquet("gs://gsod_23456/"+folder_hl+"/df_mean_wgt_HL.parquet")
# mean_wgt_HL_test_df = spark.read.parquet("gs://gsod_23456/"+folder_hl+"/test_df_mean_wgt_HL.parquet")

zero_wgt_HL_df = spark.read.parquet("gs://gsod_23456/"+folder_hl+"/df_zero_wgt_HL.parquet")
zero_wgt_HL_test_df = spark.read.parquet("gs://gsod_23456/"+folder_hl+"/test_df_zero_wgt_HL.parquet")


print(f'... completed job in {time.time() - start} seconds.')

# mean_wgt_HL_df.count(), mean_wgt_HL_test_df.count() , zero_wgt_HL_df.count(), zero_wgt_HL_test_df.count()
zero_wgt_HL_df.count(), zero_wgt_HL_test_df.count()

# # see the recoded variables
# start = time.time()
# print("Counts for Each Category:")
# for i in zero_wgt_HL_df.columns[:1]: 
zero_wgt_HL_df.groupby('_c14_wgt').count().sort(desc('count')).show()
print(f'... completed job in {time.time() - start} seconds.')

# # see the recoded variables
# start = time.time()
# print("Counts for Each Category:")
# for i in zero_wgt_HL_df.columns[:1]: 
zero_wgt_HL_test_df.groupby('_c14_wgt').count().sort(desc('count')).show()
print(f'... completed job in {time.time() - start} seconds.')

# convert nulls to L for variable columns
col = ['_c14_wgt', '_c15_wgt', '_c16_wgt', '_c17_wgt', '_c18_wgt', '_c19_wgt', '_c20_wgt', '_c21_wgt', '_c23_wgt', '_c24_wgt', '_c25_wgt', '_c26_wgt', '_c27_wgt', '_c28_wgt', '_c29_wgt', '_c30_wgt', '_c31_wgt', '_c32_wgt', '_c33_wgt', '_c34_wgt', '_c35_wgt', '_c36_wgt', '_c37_wgt', '_c38_wgt', '_c39_wgt', '_c22_wgt']
zero_wgt_HL_df = zero_wgt_HL_df.fillna('L')
zero_wgt_HL_test_df = zero_wgt_HL_test_df.fillna('L')

# confirm 
zero_wgt_HL_df.groupby('_c14_wgt').count().sort(desc('count')).show()
print(f'... completed job in {time.time() - start} seconds.')
zero_wgt_HL_test_df.groupby('_c14_wgt').count().sort(desc('count')).show()
print(f'... completed job in {time.time() - start} seconds.')
# zero_wgt_HL_df.show(2)

# broadcasting the lookup dataframe as we have to use it multiple times
lookup_df_HL = broadcast(spark.createDataFrame(
    [('H',  Vectors.dense(1.0)), 
     ('L', Vectors.dense(0.0))],
    ["id", "category"]))

# # broadcasting the lookup dataframe as we have to use it multiple times
# lookup_df_HLM = broadcast(spark.createDataFrame(
#     [('H6', Vectors.dense(1.0,0.0,0.0)), 
#      ('H1', Vectors.dense(0.0,1.0,0.0)),
#      ('L',  Vectors.dense(0.0,0.0,1.0)),
#      ('M',  Vectors.dense(0.0,0.0,0.0))],
#     ["id", "category"]))

f1_scores = {}
evaluator = MulticlassClassificationEvaluator(labelCol = "label", 
                                              predictionCol="prediction", 
                                              metricName="f1")

**<center>Algorithm Modeling Matrix</center>**

|	 3 Transformation Types	|	2 Imputing Methods	|	2 Algorithms	|	2 Model Runs	|
|	---	|	---	|	---	|	---	|
|	Weighted Value	|	Nulls => Mean	|	Logistic Regression	|	Default	|
|	Hi Low	|	Nulls => 0	|	Decision Tree	|	Hypertuned	|
|	Hi Mid Low Missing	|		|		|		|

<br>

## Logistic Regression

### Recategorized HL with Mean

# start = time.time()

# lr = LogisticRegression(featuresCol = 'features', labelCol = 'label', maxIter=10)
# lrModel = lr.fit(pipeline_prep(column_vector_maker(feature_prep(mean_wgt_HL_df)[0], 
#                                   feature_prep(mean_wgt_HL_df)[1], lookup_df_HL, "HL")))
# trainingSummary = lrModel.summary
# print("Train Results:")
# train_results(trainingSummary)

# print("\nTest Results:")
# predictions = lrModel.transform(pipeline_prep(column_vector_maker(feature_prep(mean_wgt_HL_test_df)[0], 
#                                   feature_prep(mean_wgt_HL_test_df)[1], lookup_df_HL, "HL")))
# test_results(predictions)

# f1_scores['lr_mean_wgt_HL'] = format(evaluator.evaluate(predictions), '.6f')

# print(f'... completed job in {time.time() - start} seconds.')

# # parameter tuned
# start = time.time()

# lr = LogisticRegression(featuresCol = 'features', labelCol = 'label', maxIter=10)
# lrModel = lr_cross_validate(lr).fit(pipeline_prep(column_vector_maker(feature_prep(mean_wgt_HL_df)[0], 
#                                   feature_prep(mean_wgt_HL_df)[1], lookup_df_HL, "HL")))

# print("Test Results:")
# predictions = lrModel.transform(pipeline_prep(column_vector_maker(feature_prep(mean_wgt_HL_test_df)[0], 
#                                   feature_prep(mean_wgt_HL_test_df)[1], lookup_df_HL, "HL")))
# test_results(predictions)

# f1_scores['lr_mean_wgt_HL_tuned'] = format(evaluator.evaluate(predictions), '.6f')

# print(f'... completed job in {time.time() - start} seconds.')

<br>

### Recategorized HL with Zero

start = time.time()

lr = LogisticRegression(featuresCol = 'features', labelCol = 'label', maxIter=10)
lrModel = lr.fit(pipeline_prep(column_vector_maker(feature_prep(zero_wgt_HL_df)[0], 
                                  feature_prep(zero_wgt_HL_df)[1], lookup_df_HL, "HL")))
trainingSummary = lrModel.summary
print("Train Results:")
train_results(trainingSummary)

print("\nTest Results:")
predictions = lrModel.transform(pipeline_prep(column_vector_maker(feature_prep(zero_wgt_HL_test_df)[0], 
                                  feature_prep(zero_wgt_HL_test_df)[1], lookup_df_HL, "HL")))
test_results(predictions)

f1_scores['lr_zero_wgt_HL'] = format(evaluator.evaluate(predictions), '.6f')

print(f'... completed job in {time.time() - start} seconds.')

# parameter tuned
start = time.time()

lr = LogisticRegression(featuresCol = 'features', labelCol = 'label', maxIter=10)
lrModel = lr_cross_validate(lr).fit(pipeline_prep(column_vector_maker(feature_prep(zero_wgt_HL_df)[0], 
                                  feature_prep(zero_wgt_HL_df)[1], lookup_df_HL, "HL")))

print("Test Results:")
predictions = lrModel.transform(pipeline_prep(column_vector_maker(feature_prep(zero_wgt_HL_test_df)[0], 
                                  feature_prep(zero_wgt_HL_test_df)[1], lookup_df_HL, "HL")))
test_results(predictions)

f1_scores['lr_zero_wgt_HL_tuned'] = format(evaluator.evaluate(predictions), '.6f')

print(f'... completed job in {time.time() - start} seconds.')

<br>

## Decision Tree

### Recategorized with Mean HL 

# start = time.time()

# dt = DecisionTreeClassifier(featuresCol = 'features', labelCol = 'label', maxDepth = 3)
# dtModel = dt.fit(pipeline_prep(column_vector_maker(feature_prep(mean_wgt_HL_df)[0], 
#                                   feature_prep(mean_wgt_HL_df)[1], lookup_df_HL, "HL")))

# print("Test Results:")
# predictions = dtModel.transform(pipeline_prep(column_vector_maker(feature_prep(mean_wgt_HL_test_df)[0], 
#                                   feature_prep(mean_wgt_HL_test_df)[1], lookup_df_HL, "HL")))
# test_results(predictions)

# f1_scores['dt_mean_wgt_HL'] = format(evaluator.evaluate(predictions), '.6f')

# print(f'... completed job in {time.time() - start} seconds.')

# # parameter tuned
# start = time.time()

# dt = DecisionTreeClassifier(featuresCol = 'features', labelCol = 'label', maxDepth = 3)
# dtModel = dt_cross_validate(dt).fit(pipeline_prep(column_vector_maker(feature_prep(mean_wgt_HL_df)[0], 
#                                   feature_prep(mean_wgt_HL_df)[1], lookup_df_HL, "HL")))

# print("Test Results:")
# predictions = dtModel.transform(pipeline_prep(column_vector_maker(feature_prep(mean_wgt_HL_test_df)[0], 
#                                   feature_prep(mean_wgt_HL_test_df)[1], lookup_df_HL, "HL")))
# test_results(predictions)

# f1_scores['dt_mean_wgt_HL_tuned'] = format(evaluator.evaluate(predictions), '.6f')

# print(f'... completed job in {time.time() - start} seconds.')

<br>

<br>

### Recategorized with Zero HL & HLM

start = time.time()

dt = DecisionTreeClassifier(featuresCol = 'features', labelCol = 'label', maxDepth = 3)
dtModel = dt.fit(pipeline_prep(column_vector_maker(feature_prep(zero_wgt_HL_df)[0], 
                                  feature_prep(zero_wgt_HL_df)[1], lookup_df_HL, "HL")))

print("Test Results:")
predictions = dtModel.transform(pipeline_prep(column_vector_maker(feature_prep(zero_wgt_HL_test_df)[0], 
                                  feature_prep(zero_wgt_HL_test_df)[1], lookup_df_HL, "HL")))
test_results(predictions)

f1_scores['dt_zero_wgt_HL'] = format(evaluator.evaluate(predictions), '.6f')

print(f'... completed job in {time.time() - start} seconds.')

# parameter tuned
start = time.time()

dt = DecisionTreeClassifier(featuresCol = 'features', labelCol = 'label', maxDepth = 3)
dtModel = dt_cross_validate(dt).fit(pipeline_prep(column_vector_maker(feature_prep(zero_wgt_HL_df)[0], 
                                  feature_prep(zero_wgt_HL_df)[1], lookup_df_HL, "HL")))

print("Test Results:")
predictions = dtModel.transform(pipeline_prep(column_vector_maker(feature_prep(zero_wgt_HL_test_df)[0], 
                                  feature_prep(zero_wgt_HL_test_df)[1], lookup_df_HL, "HL")))
test_results(predictions)

f1_scores['dt_zero_wgt_HL_tuned'] = format(evaluator.evaluate(predictions), '.6f')

print(f'... completed job in {time.time() - start} seconds.')

<br>

<br>

print(f'... master completed job in {time.time() - master_start} seconds.')

<br>
